move utils.astd1 and amean1
